+++ src/components/Auth/index.ts
@@ -1,2 +1,0 @@
-export { LoginForm } from "./LoginForm";
-// export { GoogleLoginButton } from "./GoogleLoginButton";